+++ tests/Phuria/SQLBuilder/Test/Unit/QueryBuilder/SelectBuilderTest.php
@@ -25,7 +25,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectMinFromTest()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->from('test');
         $qb->addSelect('MIN(test.id)');
@@ -38,7 +38,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectTableWithAlias()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('test');
         $rootTable->setAlias('SRC');
@@ -52,7 +52,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectTwoColumns()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->from('test');
         $qb->addSelect('test.id');
@@ -66,7 +66,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectColumnReferences()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('example');
         $qb->addSelect($rootTable->column('id'));
@@ -84,7 +84,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectMaxColumnReference()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('example');
         $qb->addSelect("MAX({$rootTable->column('points')})");
@@ -101,7 +101,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithWhereClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('example');
         $qb->addSelect($rootTable->column('*'));
@@ -119,7 +119,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithWhereColumnReference()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('example');
         $qb->addSelect('example.id');
@@ -135,7 +135,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithConnectedWhereClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $rootTable = $qb->from('example');
         $qb->addSelect('*');
@@ -156,7 +156,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithCrossJoinClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->from('example');
         $qb->addSelect('*');
@@ -170,7 +170,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithMultipleFromTables()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('example');
         $testTable = $qb->addFrom('test');
@@ -189,7 +189,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithLeftJoinClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('example');
         $qb->addSelect('*');
@@ -204,7 +204,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithDifferentJoins()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $userTable = $qb->from('users');
         $userTable->setAlias('u');
@@ -224,7 +224,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
 
         static::assertSame($expectedSQL, $qb->buildSQL());
 
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->from('users', 'u');
         $qb->addSelect('*');
@@ -239,7 +239,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithLimitClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->from('example');
         $qb->addSelect('*');
@@ -257,7 +257,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithGroupByClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('price_list');
         $exampleTable->setAlias('p');
@@ -273,12 +273,12 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectFromSubQuery()
     {
-        $maxQb = static::queryBuilder()->select();
+        $maxQb = static::queryBuilder()->createSelect();
 
         $exampleTable = $maxQb->from('example');
         $maxQb->addSelect("MAX({$exampleTable->column('value')}) AS max_value");
 
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
         $subQuery = $qb->from($maxQb);
         $subQuery->setAlias('SRC');
         $qb->addSelect($subQuery->column('max_value'));
@@ -297,7 +297,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithOrderByClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('example');
         $qb->addSelect('*');
@@ -317,7 +317,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectOnly()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
         $qb->addSelect('1 + 1');
 
         static::assertSame('SELECT 1 + 1', $qb->buildSQL());
@@ -328,7 +328,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithOrderedGroupBy()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('example');
         $qb->addSelect('*');
@@ -342,7 +342,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithHavingClause()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $exampleTable = $qb->from('example');
         $qb->addSelect("SUM({$exampleTable->column('price')}) AS price");
@@ -357,7 +357,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectWithGroupByRollUp()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $table = $qb->from('example');
         $qb->addSelect("SUM(".$table->column('price').")");
@@ -374,7 +374,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectFromWithAlias()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $table = $qb->from('example', 'e');
         $qb->addSelect($table->column('name'));
@@ -387,7 +387,7 @@ class SelectBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function selectMultipleArguments()
     {
-        $qb = static::queryBuilder()->select();
+        $qb = static::queryBuilder()->createSelect();
 
         $qb->addSelect('1+1', '2+2');